Refactor: split style.py and widget.py into feature-organized packages
style.py (294 lines) → style/ package:
- _artists.py   private artist helpers
- _figure.py    figure size, background, redraw
- _typography.py  font size
- _colors.py    colors, colormaps, detect_plot_type
- _opacity.py   series alpha
- _axes.py      per-axis title/labels/limits
- _legend.py    legend position/labels/bbox + LEGEND_LOCATIONS
- _grid_spines.py  grid, spines + SPINE_STYLES
- __init__.py   re-exports all public names (backward compatible)

widget.py (834 lines) → widget/ package:
- _constants.py   _PREVIEW_HEIGHT, _KNOWN_SECTIONS, _GITHUB_ISSUES
- _theme.py       _theme_css, _ios_toggle (CSS scoping + iOS toggle)
- _ctx.py         _PanelCtx dataclass (shared state across section builders)
- _helpers.py     _section, _per_btn, _lim_sliders, swatch/gradient builders
- _sections/      one file per UI section (8 files, each 25–147 lines)
- _studio.py      studio() main function (~153 lines, down from ~600)
- __init__.py     re-exports studio, available_sections

Public API and all existing imports are unchanged. All 56 tests pass.
