+++ mplstudio/widget/__init__.py
@@ -1,0 +1,5 @@
+"""ipywidgets control panel for mplstudio."""
+
+from ._studio import studio, available_sections
+
+__all__ = ["studio", "available_sections"]
