+++ mplstudio/widget/_constants.py
@@ -1,0 +1,10 @@
+"""Module-level constants for the widget package."""
+
+_PREVIEW_HEIGHT = 400
+
+_KNOWN_SECTIONS = frozenset({
+    "figure_size", "typography", "colors", "alpha",
+    "axes", "legend", "grid_spines", "palette_suggestions",
+})
+
+_GITHUB_ISSUES = "https://github.com/symoon9/mplstudio/issues"
